refactor(module:resizable): refactor
