--- components/resizable/nz-resize-handles.component.html
@@ -1,1 +1,0 @@
-<nz-resize-handle *ngFor="let dir of directions" [nzDirection]="dir"></nz-resize-handle>
